backend/pyproject.toml
[tool.poetry]
name = "scoreboard-ai-backend"
version = "0.1.0"
description = "Football match score prediction API"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
anthropic = "^0.40.0"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
